BlindStick.ino
Obstacle detection is one of the major concerns for a fully or a partially blind person (e.g. person suffering from night-blindness, cataracts, glaucoma, albinism or injury). Presented here is a smart stick using Arduino Uno. The stick uses Ultrasonic sensors for obstacle detection. The main aim of this project is to detect nearby obstacle and notify the user of the direction of that obstacle, thereby enabling the user to determine the corrective direction to head